# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BillingInvoiceList": {
        "additionalProperties": false,
        "description": "This practice's invoices, newest period first.",
        "properties": {
          "invoices": {
            "items": {
              "$ref": "#/components/schemas/BillingInvoice"
            },
            "title": "Invoices",
            "type": "array"
          },
          "total": {
            "description": "Wie viele zurückgegeben wurden. — How many were returned.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "BillingInvoiceList",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The body of every non-2xx response this API produces.\n\nDeclared as a model so it appears in the OpenAPI document and the generated TypeScript, rather\nthan being an undocumented dict that clients discover by hitting it in production.",
        "properties": {
          "detail": {
            "description": "Compatibility mirror, where FastAPI clients already look. Usually an object holding the same four fields; on a `VALIDATION_ERROR` it is Pydantic's own list of field errors, and on an error raised by a route as `HTTPException` it is whatever that route passed. Read the top-level fields instead — this one is deliberately untyped because it reproduces what shipped.",
            "title": "Detail"
          },
          "details": {
            "additionalProperties": true,
            "description": "Specifics a client can act on: line and column, unknown Ziffern, partials.",
            "title": "Details",
            "type": "object"
          },
          "error": {
            "description": "Lowercase spelling of `error_code`, for clients written before it existed.",
            "title": "Error",
            "type": "string"
          },
          "error_code": {
            "description": "Stable machine-readable code. See docs/errors.md.",
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "description": "Human-readable. May change; do not switch on it.",
            "title": "Message",
            "type": "string"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Seconds to wait before retrying the identical request. Present only when retrying could plausibly succeed; also sent as the `Retry-After` header.",
            "title": "Retry After"
          },
          "status": {
            "description": "The HTTP status, repeated in the body so a logged payload is self-contained.",
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "error_code",
          "message",
          "error",
          "status"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PlanCatalog": {
        "additionalProperties": false,
        "description": "Every plan a practice may move to, cheapest first.",
        "properties": {
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanSummary"
            },
            "title": "Plans",
            "type": "array"
          }
        },
        "title": "PlanCatalog",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AzmothApiKey": {
        "description": "Der API-Schlüssel der Praxis bzw. des Rechnungszentrums, im Format `azm_live_<id>_<secret>`. Er bestimmt zugleich die Organisation, deren Daten die Anfrage sehen darf — es gibt keinen Header, mit dem ein Aufrufer eine andere Organisation angeben könnte.\n\nThe practice's API key. It also determines which organisation's data the request may reach; there is no header with which a caller could name a different one. Mint one with `POST /api/v1/settings/api-keys` from a signed-in session.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "GOÄ coding where the legal reasoning is symbolic and auditable.\n\n`POST /api/v1/solve` takes structured clinical entities — no model runs anywhere in this service. A Soufflé Datalog program decides what is certainly chargeable, a Clingo ASP program resolves what genuinely requires a choice under a hard timeout, an independent validation pass re-checks the result, and every line carries a proof tree with rule ids and paragraph references.\n\n**The response is a DRAFT proposal, not an invoice.** It leaves `DRAFT` only when a named person approves it. Every response carries `receipt_hash` (SHA-256 over the catalog, the rule tables, the logic programs, the solver versions, the policy and the input) and the rule-coverage counts: `enforced_rule_count` is what can suppress a position, `advisory_rule_count` is what only warns.\n\nThe catalog is imported from the official GOÄ published at gesetze-im-internet.de. Rule coverage is **partial** — see `GET /api/v1/catalog`. **Synthetic data only.**\n\n**Errors.** Every non-2xx response carries `error_code` (stable, machine-readable), `message`, `details` and, where retrying could work, `retry_after` in seconds plus a `Retry-After` header. The complete catalog of codes is in `docs/errors.md`.",
    "title": "Azmoth GOÄ engine",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/billing/invoices": {
      "get": {
        "description": "Die abgeschlossenen Abrechnungszeiträume dieser Praxis, neuester zuerst.\n\nEin Eintrag entsteht, wenn ein Zeitraum endet — nicht durch einen Zahlungsvorgang. Er nennt den\nTarif, unter dem der Zeitraum berechnet wurde, das enthaltene Kontingent, die tatsächlich\ngeprüften Rechnungen, den Mehrverbrauch und die Summe. Alle Beträge in Euro-Cent.\n\n**Ein Zeitraum ohne Verbrauch erhält ebenfalls einen Eintrag**, mit Summe 0. „August: nichts\nfällig\" und „August: kein Eintrag\" sind verschiedene Aussagen, und nur die erste lässt sich\nabgleichen.\n\n**Dies ist keine Rechnung im rechtlichen Sinn.** Keine Umsatzsteuer, keine Rechnungsnummer im\nSinne des § 14 UStG, keine Zahlungsbedingungen — es ist die Grundlage, aus der eine solche\nRechnung erstellt wird.\n\nDer laufende, noch nicht abgeschlossene Zeitraum steht in `GET /api/v1/billing/usage` und\nerscheint hier bewusst nicht: ein Betrag, der sich noch ändert, gehört nicht in eine Liste\nabgerechneter Zeiträume.\n\n---\n\nThe closed billing periods, newest first — one row per period that ended, with the plan it was\npriced under and every amount in euro cents. A period with no usage still gets a row, with a\nzero total. The open period is not listed; it is in `GET /api/v1/billing/usage`.\n\nNot a Rechnung in the legal sense: no VAT, no invoice-number series, no payment terms.",
        "operationId": "read_invoices_api_v1_billing_invoices_get",
        "parameters": [
          {
            "description": "Wie viele Zeiträume höchstens, neueste zuerst. — At most this many periods, newest first.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Wie viele Zeiträume höchstens, neueste zuerst. — At most this many periods, newest first.",
              "maximum": 120,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingInvoiceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          }
        },
        "security": [
          {
            "AzmothApiKey": []
          }
        ],
        "summary": "Abgerechnete Zeiträume",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/plans": {
      "get": {
        "description": "Die Tarife, auf die diese Praxis wechseln kann — vom günstigsten zum teuersten.\n\nAbgelöste Revisionen erscheinen hier **nicht**. Preise werden nie geändert, sondern durch eine\nneue Revision ersetzt; eine Praxis auf einer älteren Revision behält deren Konditionen, aber\nneu wählbar ist nur die aktuelle. Der eigene Tarif steht in `GET /api/v1/billing/usage` und\nkann eine abgelöste Revision sein.\n\n**Alle Beträge in Euro-Cent als ganze Zahl.** `9900` bedeutet 99,00 €.\n\n---\n\nThe plans a practice may move to, cheapest first. Superseded revisions are excluded: prices are\nappend-only, so a practice on an older revision keeps its terms but only the current one can be\nnewly selected — their own plan is in `GET /api/v1/billing/usage` and may be a superseded code.\n\nRequires a credential (a key or a session) even though the catalog is not tenant data. That is a\ndeliberate choice rather than an oversight: a price list is commercially sensitive before launch,\nand the marketing site is where a public one belongs.",
        "operationId": "read_plans_api_v1_billing_plans_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          }
        },
        "security": [
          {
            "AzmothApiKey": []
          }
        ],
        "summary": "Verfügbare Tarife",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/upgrade": {
      "post": {
        "description": "Wechselt den Tarif dieser Praxis. Nur aus einer angemeldeten Sitzung, nicht mit API-Schlüssel.\n\nZieltarif entweder als `tier` (`free`, `starter`, `pro`, `enterprise` — löst auf die aktuell\nwählbare Revision auf) oder als genaue `plan_code`. Genau eine der beiden Angaben ist\nerforderlich.\n\n**Der Abrechnungszeitraum wird nicht neu gestartet.** Ein Wechsel mitten im Zeitraum erhöht das\nKontingent und lässt das Fenster unverändert: wer 480 von 500 Rechnungen verbraucht hat und auf\nPro wechselt, hat sofort 2.020 übrig — nicht ein neues Kontingent und eine Zeitraumgrenze mitten\nim Monat. Eine anteilige Berechnung der Grundgebühr (Proration) ist bewusst nicht implementiert;\nsiehe `docs/BILLING.md`.\n\n**Auch ein Wechsel nach unten ist erlaubt.** Der Aufruf heisst `upgrade`, weil das der Normalfall\nist; die Engine verweigert keinen Wechsel nach unten. Liegt der bisherige Verbrauch über dem\nneuen Kontingent, gilt ab sofort das neue: bei einem Tarif ohne Mehrverbrauch führt das zu\n`429 QUOTA_EXCEEDED`, bis der Zeitraum wechselt. `changed: false` heisst, die Praxis war bereits\nauf diesem Tarif — ein erfolgreiches No-op und kein Fehler.\n\nFehler: `422` wenn weder `tier` noch `plan_code` angegeben ist oder beide, `404` wenn der Tarif\nnicht existiert oder nicht (mehr) wählbar ist, `403` wenn die Anfrage keine Praxis nennt.\n\n---\n\nChange this practice's plan. Session only — not with an API key, because it is a commercial\ncommitment and a key must not be able to escalate its own entitlements. See the module docstring.\n\nTarget named by `tier` or by exact `plan_code`; exactly one. The period is not restarted, and a\ndowngrade is permitted. Idempotent in the sense that matters: naming the current plan succeeds\nand reports `changed: false`.",
        "operationId": "upgrade_subscription_api_v1_billing_upgrade_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          }
        },
        "summary": "Tarif wechseln",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/usage": {
      "get": {
        "description": "Tarif, verbrauchte Rechnungen, verbleibendes Kontingent und angefallener Mehrverbrauch.\n\n**Gezählt werden Rechnungen, nicht Aufrufe.** Ein Bulk-Upload mit 300 Lieferungen ist ein\nAPI-Aufruf und 300 Rechnungen — `invoices_processed` ist die abrechenbare Einheit,\n`requests` steht nur zur Einordnung daneben.\n\nDer Zeitraum ist der Abrechnungszeitraum dieser Praxis und **nicht** der Kalendermonat: er ist\nan den Tag gebunden, an dem die Praxis angelegt wurde. `GET /api/v1/settings/usage` beantwortet\nbewusst die andere Frage — Verbrauch im laufenden Kalendermonat — und beide nennen den Zeitraum,\nden sie verwendet haben. Wer beide Zahlen vergleicht, muss auf diese Fenster achten.\n\n`projected_total_cents` ist keine Prognose, sondern Grundgebühr plus bisher angefallener\nMehrverbrauch: eine Zahl, die nur steigt und die zwei Leser gleich lesen.\n\n---\n\nThe open billing period: which plan, how many invoices audited, how many remain, what overage\nhas accrued. Readable with an API key or from a session; the credential decides the practice.\n\nThe window is the practice's own billing period, anchored on the day they were created — not the\ncalendar month, which is what `GET /api/v1/settings/usage` reports. Both state their window.\n\nReading this also **rolls the period** if the stored one has ended, which writes an invoice for\neach period that closed. There is no scheduler in this service, so the first request of a new\nperiod is what closes the previous one — see `app.services.billing`.",
        "operationId": "read_billing_usage_api_v1_billing_usage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingUsage"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "See docs/errors.md for the codes."
          }
        },
        "security": [
          {
            "AzmothApiKey": []
          }
        ],
        "summary": "Kontingent und Verbrauch",
        "tags": [
          "billing"
        ]
      }
    },
...
  },
...
}